Desktop/projects/wifi_speed_test/requirements.txt
 
speedtest-cli
